+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/BulkProcessingOptions.java
@@ -3,7 +3,10 @@
 
 package com.azure.cosmos;
 
+import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
 import com.azure.cosmos.implementation.batch.BatchRequestResponseConstants;
+import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
+import com.azure.cosmos.models.CosmosItemRequestOptions;
 import com.azure.cosmos.util.Beta;
 import reactor.core.publisher.Flux;
 
@@ -24,6 +27,7 @@ public final class BulkProcessingOptions<TContext> {
         BatchRequestResponseConstants.DEFAULT_MAX_MICRO_BATCH_INTERVAL_IN_MILLISECONDS);
     private final TContext batchContext;
     private final BulkProcessingThresholds<TContext> thresholds;
+    private OperationContextAndListenerTuple operationContextAndListenerTuple;
 
     @Beta(value = Beta.SinceVersion.V4_17_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public BulkProcessingOptions(TContext batchContext, BulkProcessingThresholds<TContext> thresholds) {
@@ -153,5 +157,32 @@ public final class BulkProcessingOptions<TContext> {
     @Beta(value = Beta.SinceVersion.V4_17_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public BulkProcessingThresholds<TContext> getThresholds() {
         return this.thresholds;
+    OperationContextAndListenerTuple getOperationContextAndListenerTuple() {
+        return this.operationContextAndListenerTuple;
+    }
+
+    void setOperationContextAndListenerTuple(OperationContextAndListenerTuple operationContextAndListenerTuple) {
+        this.operationContextAndListenerTuple = operationContextAndListenerTuple;
+    }
+
+    ///////////////////////////////////////////////////////////////////////////////////////////
+    // the following helper/accessor only helps to access this class outside of this package.//
+    ///////////////////////////////////////////////////////////////////////////////////////////
+
+    static {
+        ImplementationBridgeHelpers.CosmosBulkProcessingOptionsHelper.setCosmosBulkProcessingOptionAccessor(
+            new ImplementationBridgeHelpers.CosmosBulkProcessingOptionsHelper.CosmosBulkProcessingOptionAccessor() {
+
+                @Override
+                public <T> void setOperationContext(BulkProcessingOptions<T> bulkProcessingOptions,
+                                                    OperationContextAndListenerTuple operationContextAndListenerTuple) {
+                    bulkProcessingOptions.setOperationContextAndListenerTuple(operationContextAndListenerTuple);
+                }
+
+                @Override
+                public <T> OperationContextAndListenerTuple getOperationContext(BulkProcessingOptions<T> bulkProcessingOptions) {
+                    return bulkProcessingOptions.getOperationContextAndListenerTuple();
+                }
+            });
     }
 }