+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/BulkExecutor.java
@@ -17,6 +17,11 @@ import com.azure.cosmos.TransactionalBatchResponse;
 import com.azure.cosmos.implementation.AsyncDocumentClient;
 import com.azure.cosmos.implementation.CosmosDaemonThreadFactory;
 import com.azure.cosmos.implementation.apachecommons.lang.tuple.Pair;
+import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
+import com.azure.cosmos.implementation.RequestOptions;
+import com.azure.cosmos.implementation.apachecommons.lang.tuple.Pair;
+import com.azure.cosmos.implementation.spark.OperationContextAndListenerTuple;
+import io.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.Exceptions;
@@ -70,6 +75,7 @@ public final class BulkExecutor<TContext> {
 
     private final CosmosAsyncContainer container;
     private final AsyncDocumentClient docClientWrapper;
+    private final OperationContextAndListenerTuple operationListener;
     private final ThrottlingRetryOptions throttlingRetryOptions;
     private final Flux<CosmosItemOperation> inputOperations;
 
@@ -108,6 +114,9 @@ public final class BulkExecutor<TContext> {
         maxMicroBatchIntervalInMs = bulkOptions.getMaxMicroBatchInterval().toMillis();
         batchContext = bulkOptions.getBatchContext();
         this.partitionScopeThresholds = BridgeInternal.getPartitionScopeThresholds(bulkOptions.getThresholds());
+        operationListener = ImplementationBridgeHelpers.CosmosBulkProcessingOptionsHelper
+            .getCosmosBulkProcessingOptionAccessor()
+            .getOperationContext(bulkOptions);
 
         // Initialize sink for handling gone error.
         mainSourceCompleted = new AtomicBoolean(false);
@@ -400,8 +409,11 @@ public final class BulkExecutor<TContext> {
     }
 
     private Mono<TransactionalBatchResponse> executeBatchRequest(PartitionKeyRangeServerBatchRequest serverRequest) {
+        RequestOptions options = new RequestOptions();
+        options.setOperationContextAndListenerTuple(operationListener);
+
         return this.docClientWrapper.executeBatchRequest(
-            BridgeInternal.getLink(this.container), serverRequest, null, false);
+            BridgeInternal.getLink(this.container), serverRequest, options, false);
     }
 
     private void completeAllSinks() {