+++ sdk/keyvault/azure-security-keyvault-keys/src/main/java/com/azure/security/keyvault/keys/cryptography/RsaKeyCryptographyClient.java
@@ -6,13 +6,15 @@ package com.azure.security.keyvault.keys.cryptography;
 import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
+import com.azure.security.keyvault.keys.cryptography.models.DecryptParameters;
 import com.azure.security.keyvault.keys.cryptography.models.DecryptResult;
-import com.azure.security.keyvault.keys.cryptography.models.EncryptionAlgorithm;
+import com.azure.security.keyvault.keys.cryptography.models.EncryptParameters;
 import com.azure.security.keyvault.keys.cryptography.models.EncryptResult;
-import com.azure.security.keyvault.keys.cryptography.models.UnwrapResult;
+import com.azure.security.keyvault.keys.cryptography.models.EncryptionAlgorithm;
 import com.azure.security.keyvault.keys.cryptography.models.KeyWrapAlgorithm;
-import com.azure.security.keyvault.keys.cryptography.models.SignatureAlgorithm;
 import com.azure.security.keyvault.keys.cryptography.models.SignResult;
+import com.azure.security.keyvault.keys.cryptography.models.SignatureAlgorithm;
+import com.azure.security.keyvault.keys.cryptography.models.UnwrapResult;
 import com.azure.security.keyvault.keys.cryptography.models.VerifyResult;
 import com.azure.security.keyvault.keys.cryptography.models.WrapResult;
 import com.azure.security.keyvault.keys.models.JsonWebKey;
@@ -55,20 +57,33 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
     }
 
     @Override
+    Mono<EncryptResult> encryptAsync(EncryptionAlgorithm algorithm, byte[] plaintext, Context context,
+                                     JsonWebKey jsonWebKey) {
+        Objects.requireNonNull(algorithm, "'algorithm' cannot be null.");
+        Objects.requireNonNull(plaintext, "'plaintext' cannot be null.");
+
+        return encryptInternal(algorithm, plaintext, context, jsonWebKey);
+    }
+
+    @Override
     Mono<EncryptResult> encryptAsync(EncryptParameters encryptParameters, Context context, JsonWebKey jsonWebKey) {
         Objects.requireNonNull(encryptParameters, "'encryptParameters' cannot be null.");
         Objects.requireNonNull(encryptParameters.getAlgorithm(), "encryptParameters.getAlgorithm() cannot be null.");
         Objects.requireNonNull(encryptParameters.getPlainText(), "encryptParameters.getPlainText() cannot be null.");
 
+        return encryptInternal(encryptParameters.getAlgorithm(), encryptParameters.getPlainText(), context, jsonWebKey);
+    }
+
+    private Mono<EncryptResult> encryptInternal(EncryptionAlgorithm algorithm, byte[] plaintext, Context context,
+                                                JsonWebKey jsonWebKey) {
         keyPair = getKeyPair(jsonWebKey);
 
         // Interpret the requested algorithm
-        EncryptionAlgorithm algorithm = encryptParameters.getAlgorithm();
         Algorithm baseAlgorithm = AlgorithmResolver.DEFAULT.get(algorithm.toString());
 
         if (baseAlgorithm == null) {
             if (serviceCryptoAvailable()) {
-                return serviceClient.encrypt(encryptParameters, context);
+                return serviceClient.encrypt(algorithm, plaintext, context);
             }
 
             return Mono.error(new NoSuchAlgorithmException(algorithm.toString()));
@@ -78,7 +93,7 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
 
         if (keyPair.getPublic() == null) {
             if (serviceCryptoAvailable()) {
-                return serviceClient.encrypt(encryptParameters, context);
+                return serviceClient.encrypt(algorithm, plaintext, context);
             }
 
             return Mono.error(new IllegalArgumentException(
@@ -92,7 +107,7 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
         try {
             transform = algo.createEncryptor(keyPair);
 
-            return Mono.just(new EncryptResult(transform.doFinal(encryptParameters.getPlainText()), algorithm,
+            return Mono.just(new EncryptResult(transform.doFinal(plaintext), algorithm,
                 jsonWebKey.getId()));
         } catch (InvalidKeyException
             | NoSuchAlgorithmException
@@ -105,20 +120,34 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
     }
 
     @Override
+    Mono<DecryptResult> decryptAsync(EncryptionAlgorithm algorithm, byte[] ciphertext, Context context,
+                                     JsonWebKey jsonWebKey) {
+        Objects.requireNonNull(algorithm, "'algorithm' cannot be null.");
+        Objects.requireNonNull(ciphertext, "'ciphertext' cannot be null.");
+
+        return decryptInternal(algorithm, ciphertext, context, jsonWebKey);
+    }
+
+    @Override
     Mono<DecryptResult> decryptAsync(DecryptParameters decryptParameters, Context context, JsonWebKey jsonWebKey) {
         Objects.requireNonNull(decryptParameters, "'decryptOptions' cannot be null.");
         Objects.requireNonNull(decryptParameters.getAlgorithm(), "decryptParameters.getAlgorithm() cannot be null.");
         Objects.requireNonNull(decryptParameters.getCipherText(), "decryptParameters.getCipherText() cannot be null.");
 
+        return decryptInternal(decryptParameters.getAlgorithm(), decryptParameters.getCipherText(), context,
+            jsonWebKey);
+    }
+
+    Mono<DecryptResult> decryptInternal(EncryptionAlgorithm algorithm, byte[] ciphertext, Context context,
+                                        JsonWebKey jsonWebKey) {
         keyPair = getKeyPair(jsonWebKey);
 
         // Interpret the requested algorithm
-        EncryptionAlgorithm algorithm = decryptParameters.getAlgorithm();
         Algorithm baseAlgorithm = AlgorithmResolver.DEFAULT.get(algorithm.toString());
 
         if (baseAlgorithm == null) {
             if (serviceCryptoAvailable()) {
-                return serviceClient.decrypt(decryptParameters, context);
+                return serviceClient.decrypt(algorithm, ciphertext, context);
             }
 
             return Mono.error(new NoSuchAlgorithmException(algorithm.toString()));
@@ -128,7 +157,7 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
 
         if (keyPair.getPrivate() == null) {
             if (serviceCryptoAvailable()) {
-                return serviceClient.decrypt(decryptParameters, context);
+                return serviceClient.decrypt(algorithm, ciphertext, context);
             }
 
             return Mono.error(new IllegalArgumentException(
@@ -142,7 +171,7 @@ class RsaKeyCryptographyClient extends LocalKeyCryptographyClient {
         try {
             transform = algo.createDecryptor(keyPair);
 
-            return Mono.just(new DecryptResult(transform.doFinal(decryptParameters.getCipherText()), algorithm,
+            return Mono.just(new DecryptResult(transform.doFinal(ciphertext), algorithm,
                 jsonWebKey.getId()));
         } catch (InvalidKeyException
             | NoSuchAlgorithmException