+++ sdk/resourcemanager/azure-resourcemanager-network/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
-## 2.6.0-beta.1 (Unreleased)
+## 2.7.0-beta.1 (Unreleased)
+
+
+## 2.6.0 (2021-06-18)
 
 - Updated `api-version` to `2021-02-01`
 - Supported multiple `ApplicationSecurityGroup` in rules of `NetworkSecurityGroup`.